+++ gmail/gmail_tools.py
@@ -21,7 +21,6 @@ from email.policy import SMTP
 from email.utils import formataddr
 
 import httpx
-from fastmcp.exceptions import ToolError as ToolExecutionError
 from mcp.types import ToolAnnotations
 
 from pydantic import Field
@@ -50,37 +49,19 @@ from auth.scopes import (
     GMAIL_MODIFY_SCOPE,
     GMAIL_LABELS_SCOPE,
 )
-from gmail.gmail_helpers import _analyze_thread_ownership_impl
+from gmail.gmail_helpers import (
+    GMAIL_METADATA_HEADERS,
+    RAW_BODY_TRUNCATE_LIMIT,
+    _analyze_thread_ownership_impl,
+    _is_benign_signature_http_error,
+    _signature_fetch_tool_error,
+)
 
 logger = logging.getLogger(__name__)
 
 GMAIL_BATCH_SIZE = 25
 GMAIL_REQUEST_DELAY = 0.1
 HTML_BODY_TRUNCATE_LIMIT = 20000
-RAW_BODY_TRUNCATE_LIMIT = 20000
-GMAIL_QUOTA_ERROR_MARKERS = (
-    "dailyLimitExceeded",
-    "quotaExceeded",
-    "rateLimitExceeded",
-    "userRateLimitExceeded",
-    "usageLimits",
-    "quota",
-    "rate limit",
-)
-
-GMAIL_METADATA_HEADERS = [
-    "Subject",
-    "From",
-    "To",
-    "Cc",
-    "Message-ID",
-    "In-Reply-To",
-    "References",
-    "Date",
-    "List-Unsubscribe",
-    "Precedence",
-    "List-Id",
-]
 LOW_VALUE_TEXT_PLACEHOLDERS = (
     "your client does not support html",
     "view this email in your browser",
@@ -425,35 +406,6 @@ def _append_signature_to_body(
         return body
     separator = "\n\n" if body.strip() else ""
     return f"{body}{separator}{signature_text}"
-
-
-def _http_error_status(error: HttpError) -> Optional[int]:
-    status = getattr(getattr(error, "resp", None), "status", None)
-    try:
-        return int(status)
-    except (TypeError, ValueError):
-        return None
-
-
-def _is_quota_or_rate_limit_error(error: HttpError) -> bool:
-    details = str(error).lower()
-    content = getattr(error, "content", None)
-    if isinstance(content, bytes):
-        details = f"{details} {content.decode('utf-8', errors='ignore').lower()}"
-    elif content:
-        details = f"{details} {str(content).lower()}"
-    return any(marker.lower() in details for marker in GMAIL_QUOTA_ERROR_MARKERS)
-
-
-def _is_benign_signature_http_error(error: HttpError) -> bool:
-    status = _http_error_status(error)
-    return status == 401 or (
-        status == 403 and not _is_quota_or_rate_limit_error(error)
-    )
-
-
-def _signature_fetch_tool_error(error: Exception) -> ToolExecutionError:
-    return ToolExecutionError(f"Failed to fetch Gmail send-as signatures: {error}")
 
 
 async def _fetch_original_for_quote(
